requirements.txt
# Core dependencies
torch==2.0.1
transformers==4.28.1
torch-geometric==2.3.0
flash-attn==2.3.3

# Scientific computing
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
h5py==3.8.0
biopython==1.81

# Data handling
datasets==2.12.0
pyarrow==12.0.0

# Experiment tracking
wandb==0.15.0
tensorboard==2.12.0

# Utilities
tqdm==4.65.0
pyyaml==6.0
omegaconf==2.3.0

# Development
jupyter==1.0.0
pytest==7.3.1
black==23.3.0
flake8==6.0.0